backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic-settings==2.3.4
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
asyncpg==0.29.0
aiosqlite==0.20.0
bcrypt==4.1.2
python-jose[cryptography]==3.5.0
email-validator==2.2.0
python-multipart==0.0.21
httpx==0.27.0
pytest==8.2.2
stripe==11.0.0
Pillow==10.4.0
reportlab==4.4.9
sentry-sdk==2.48.0
